debugging

+++ train.py
@@ -102,7 +102,7 @@ def train(nn_model,train_set_loader,val_set_loader,loss_func,optimizer, config):
                 image,gt = image, gt
             
             
-            output = nn_model(image)
+            output = nn_model(image.float())
             loss = loss_func(output, gt)
